refactoring and first test
