+++ vcs/src/main/java/vcs/commands/CheckoutCommand.java
@@ -2,8 +2,8 @@ package vcs.commands;
 
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.Parameters;
-import vcs.CommitConfig;
-import vcs.GlobalConfig;
+import vcs.config.CommitConfig;
+import vcs.config.GlobalConfig;
 import vcs.util.VcsUtils;
 
 import java.io.IOException;
