+++ vcs/src/main/java/vcs/config/CommitConfig.java
@@ -1,4 +1,4 @@
-package vcs;
+package vcs.config;
 
 import vcs.util.VcsUtils;
