+++ vcs/src/test/java/VcsTest.java
@@ -1,0 +1,14 @@
+import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
+
+public class VcsTest {
+
+    @Test
+    public void simpleTest() {
+        VCS.run("init");
+
+        assertEquals(true, true);
+    }
+
+}
